Change the model with `Section`
Currently, we explicit each part of a Food (description, preservations, preparations, recipes, replacements).
When creating a Food, we directly have all these elements at the same time and we cannot reorder them.

We change the model to have the notion of `Section` and a `Food` is only made of `Section`s.

With this design, we give more freedom in the structure without removing the semantic.