Change the logic of the creation and update of Food
Currently, we create a food with the whole form.
Then if we want to update it, we return in the same big form.

This last one is quite complex and not quite interesting.

We change the UX to be able to update the Food directly in the view page.

This means changing lot of logic and removing the update page.