Change the format of `InputDuration`
Currently, the `InputDuration` uses an input number in hours and minutes.
When we want to use it for greater durations, it is not adapted anymore.

We change it to an input range with a list of acceptable durations.
With this strategy, the caller can specify which durations it accepts.

We also take this opportunity to use ISO-8601 for duration format.