Update the design of the menu
We update the design of the menu so it becomes more intuitive to use it.

It impacts the use of `CategoriesPage` and this last one becomes useless. We remove it and replace it by `CategoryPage`.
For now, this last one is quite empty but could be filled afterward (with nested foods or description or anything else).

Also, we take this opportunity to correct sort with accents.