Restructure behavior of Food pages
Currently, we split routes between food and category.

We change the structure of the application so the routes include both information.

It gives us the capacity to correctly select the right items (category and food) in the menu.

Also, we update the model so a category is now a required attribute in Food.

We change the behavior of all Food pages by using the category directly in the url. But we still give to the user the opportunity to change the category.