Correct behavior of page change
When changing page, but in the same component, the page was not refreshed.

There was these two issues:

* No transition
* Some part of the page is not updated

The reason was because Svelte reused the exact same component (so no mount / destroy and no animation on component creation / destruction).

To correctly update the page, we change the `Page` component by adding a keyed-each block (see https://github.com/sveltejs/svelte/issues/1469).

There was some issue with this solution because each time we update the component, the render is updated so the Page and create an infinite loop.
To avoid this problem, we remove the call to `Page` component from concrete pages and include it directly with the Router. To do so, we create the PageRoute component behaving as expected.