Move searches on the right
Move searches on the right instead of the bottom of the `FoodPage`

Also, we initially hide searches and make them visible on demand instead of always showing the full list.

Then, we move the management of searches directly in this list.
So we can also remove the page managing them currently.